+++ slim/base/types/inner_interface_name.py
@@ -5,4 +5,5 @@ class InnerInterfaceName:
     LIST_WITH_SIZE = 'list_size'
     SET = 'set'
     NEW = 'new'
+    BULK_INSERT = 'bulk_insert'
     DELETE = 'delete'