+++ slim/ext/openapi/serve.py
@@ -41,7 +41,8 @@ def doc_serve(app: 'Application'):
 
         options = ''
         for i in app.permission.roles:
-            options += '<option value=%r>%s</option>' % (get_query_by_role(i), i or 'visitor')
+            selected = 'selected ' if role == i else ''
+            options += '<option ' + selected + 'value=%r>%s</option>' % (get_query_by_role(i), i or 'visitor')
 
         change_role_html = '''
 <div id="change-role" style="position: fixed; top: 10px; right: 10px; z-index: 100; display: flex; align-items: center;background: white;padding: 4px;border-radius: 1px;">