+++ tests/support_tests/test_peewee.py
@@ -2,7 +2,6 @@ import json
 
 import pytest
 from unittest import mock
-from aiohttp.test_utils import make_mocked_request
 from multidict import MultiDict
 
 from slim.base.user import BaseUserViewMixin
@@ -10,6 +9,7 @@ from slim.retcode import RETCODE
 from slim.support.peewee import PeeweeView
 from peewee import *
 from slim import Application, ALL_PERMISSION
+from slim.tools.test import invoke_interface
 
 pytestmark = [pytest.mark.asyncio]
 app = Application(cookies_secret=b'123456', permission=ALL_PERMISSION)
@@ -31,40 +31,22 @@ class ATestView(PeeweeView):
 
 db.create_tables([ATestModel])
 
+app._prepare()
+
 
 async def test_post_blob():
-    request = make_mocked_request('POST', '/api/test', headers={},
-                                  protocol=mock.Mock(), app=app)
-    view = ATestView(app, request)
-    await view._prepare()
-    await view.get()
+    view = await invoke_interface(app, ATestView().get, json_request=False)
     assert view.ret_val['code'] == RETCODE.NOT_FOUND
 
-    request._post = dict(info='aabbcc')
-    view = ATestView(app, request)
-    await view._prepare()
-    await view.new()
+    view = await invoke_interface(app, ATestView().new, post={'info': 'aabbcc'}, json_request=False)
     assert view.ret_val['code'] == RETCODE.SUCCESS
 
-    request._post = dict(info='a')  # 0x0A
-    view = ATestView(app, request)
-    await view._prepare()
-    await view.new()
+    view = await invoke_interface(app, ATestView().new, post={'info': 'a'}, json_request=False)  # 0x0A
     assert view.ret_val['code'] == RETCODE.SUCCESS
 
-    view = ATestView(app, request)
-    await view._prepare()
-    await view.get()
+    view = await invoke_interface(app, ATestView().get)
     assert view.ret_val['code'] == RETCODE.SUCCESS
     assert view.ret_val['data']['info'] == b'\xaa\xbb\xcc'
 
-    request = make_mocked_request('POST', '/api/test',
-                                  headers={'content-type': 'application/json'},
-                                  protocol=mock.Mock(), app=app)
-    raw_json = json.dumps({'info': 'aabbcc'})
-    request._post = raw_json
-    request._read_bytes = bytes(raw_json, encoding='utf-8')
-    view = ATestView(app, request)
-    await view._prepare()
-    await view.new()
+    view = await invoke_interface(app, ATestView().new, post={'info': 'aabbcc'})
     assert view.ret_val['code'] == RETCODE.SUCCESS